refactor: Split code into modules and fix typos in Girolle's speed and concurrency description
